Crime data in LA
Analytic Goals: used clustering method to analyze and visualize crime report dataset for user knowing how safety in different areas-Los Angeles. It clusters the total number of crimes occur in one area and shows it by region on the map. The heatmap and folium map is interactive, it will highlight the areas with high regional crime rate.